Fix window size and resize

+++ responder.go
@@ -6,7 +6,7 @@ package engi
 
 type Responder interface {
 	Render()
-	Resize(width, height int)
+	Resize(width, height float32)
 	Preload()
 	Setup()
 	Close()
@@ -24,7 +24,7 @@ func (g *Game) Setup()                            {}
 func (g *Game) Close()                            {}
 func (g *Game) Update(dt float32)                 {}
 func (g *Game) Render()                           {}
-func (g *Game) Resize(w, h int)                   {}
+func (g *Game) Resize(w, h float32)               {}
 func (g *Game) Mouse(x, y float32, action Action) {}
 func (g *Game) Scroll(amount float32)             {}
 func (g *Game) Key(key Key, modifier Modifier, action Action) {
